+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Added
 
 - Added RecordError to SpanOption. (#23)
+- Added DisableQuery to SpanOption. (#26)
 
 ## [0.5.0] - 2021-08-02